requirements.txt
# AI Integration
google-genai>=0.3.0
python-dotenv>=1.0.0

# Backend
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Frontend
streamlit>=1.24.0
requests>=2.31.0

# Testing
pytest>=7.0

